# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Approval System API
  version: 1.0.0
  description: 基于当前代码实现整理的接口文档（Spring Boot + Flowable）
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
  /api/workflow/requests:
    post:
      tags: [Workflow]
      summary: 发起流程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProcessRequest'
            example:
              businessKey: BR-20260224-0001
              title: 差旅申请-上海
              applicantId: 1001
              applicantDeptId: 10
              applicantPostId: 101
              processKey: approvalWorkflow
              variables:
                amount: 3200
                reason: 客户拜访
              countersignUsers: ["1002", "1003"]
              countersignMode: ALL
              passRatio: 1
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartProcessResponse'
              example:
                processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                message: Process started successfully
        '403':
          description: 权限不足（非 EMPLOYEE）
  /api/workflow/tasks:
    get:
      tags: [Workflow]
      summary: 查询用户任务
      parameters:
        - name: assignee
          in: query
          required: true
          description: 任务办理人（用户ID字符串）
          schema:
            type: string
            example: "1002"
        - name: includeCandidate
          in: query
          required: false
          description: 是否同时查询候选任务
          schema:
            type: boolean
            default: false
            example: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskInfo'
              example:
                - taskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  taskName: 部门主管审批
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  assignee: "1002"
                  createTime: "2026-02-24T10:20:30Z"
  /api/workflow/tasks/{taskId}/claim:
    post:
      tags: [Workflow]
      summary: 认领任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimTaskRequest'
            example:
              userId: "1002"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task claimed
  /api/workflow/tasks/{taskId}/complete:
    post:
      tags: [Workflow]
      summary: 完成审批任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
            example:
              userId: "1002"
              approvalResult: APPROVE
              comments: 同意，预算合理
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task completed
  /api/workflow/tasks/{taskId}/delegate:
    post:
      tags: [Workflow]
      summary: 委派任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegateTaskRequest'
            example:
              userId: "1002"
              delegateUserId: "1005"
              comment: 我出差中，委派处理
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task delegated
  /api/workflow/tasks/{taskId}/resolve:
    post:
      tags: [Workflow]
      summary: 委派处理完成（resolve）
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveTaskRequest'
            example:
              userId: "1005"
              approvalResult: APPROVE
              comment: 已核实，同意
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task resolved
  /api/workflow/tasks/{taskId}/reassign:
    post:
      tags: [Workflow]
      summary: 转办任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReassignTaskRequest'
            example:
              userId: "1"
              newAssigneeId: "1006"
              comment: 按组织调整转办
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task reassigned
  /api/workflow/tasks/{taskId}/return:
    post:
      tags: [Workflow]
      summary: 回退到会签节点
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnTaskRequest'
            example:
              userId: "1002"
              comment: 信息不完整，退回补充
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task returned
  /api/workflow/tasks/{taskId}/return/previous:
    post:
      tags: [Workflow]
      summary: 回退到上一步
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnTaskRequest'
            example:
              userId: "1002"
              comment: 退回上一步重新审批
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task returned to previous
  /api/workflow/tasks/{taskId}/return/target:
    post:
      tags: [Workflow]
      summary: 回退到指定节点
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnToTargetRequest'
            example:
              userId: "1002"
              targetActivityId: managerApproveTask
              comment: 回退到部门主管节点
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task returned to target
  /api/workflow/tasks/{taskId}/return/applicant:
    post:
      tags: [Workflow]
      summary: 回退到发起人
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnTaskRequest'
            example:
              userId: "1002"
              comment: 退回申请人修改
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Task returned to applicant
  /api/workflow/process/{processInstanceId}/cancel:
    post:
      tags: [Workflow]
      summary: 撤销流程
      parameters:
        - name: processInstanceId
          in: path
          required: true
          description: 流程实例ID
          schema:
            type: string
            example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelProcessRequest'
            example:
              userId: "1001"
              comment: 申请内容有误，撤销后重提
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
              example:
                success: true
                message: Process cancelled
  /api/requests:
    get:
      tags: [Requests]
      summary: 按数据权限查询申请单
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BizRequest'
              example:
                - id: 1
                  businessKey: BR-20260224-0001
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  processDefinitionId: approvalWorkflow:3:7504
                  formInstanceId: 88
                  applicantId: 1001
                  applicantDeptId: 10
                  applicantPostId: 101
                  title: 差旅申请-上海
                  status: 2
                  currentTaskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  currentAssigneeId: 1002
                  submitTime: "2026-02-24T10:20:30Z"
                  finishTime: null
                  createdAt: "2026-02-24T10:20:00Z"
                  updatedAt: "2026-02-24T10:30:00Z"
                  isDeleted: 0
  /api/requests/tasks:
    get:
      tags: [Requests]
      summary: 查询可见流程任务
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskInfo'
              example:
                - taskId: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
                  taskName: 部门主管审批
                  processInstanceId: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
                  assignee: "1002"
                  createTime: "2026-02-24T10:20:30Z"
...
components:
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      description: Flowable任务ID
      schema:
        type: string
        example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
    UserIdQuery:
      name: userId
      in: query
      required: true
      description: 当前登录用户ID
      schema:
        type: integer
        format: int64
        example: 1001
  schemas:
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        message:
          type: string
          description: 操作结果说明
          example: Task completed
    StartProcessRequest:
      type: object
      required: [title, applicantId]
      properties:
        businessKey:
          type: string
          description: 业务主键，不传则后端自动生成
          example: BR-20260224-0001
        title:
          type: string
          description: 申请标题
          example: 差旅申请-上海
        applicantId:
          type: integer
          format: int64
          description: 申请人用户ID
          example: 1001
        applicantDeptId:
          type: integer
          format: int64
          nullable: true
          description: 申请人部门ID
          example: 10
        applicantPostId:
          type: integer
          format: int64
          nullable: true
          description: 申请人岗位ID
          example: 101
        formInstanceId:
          type: integer
          format: int64
          nullable: true
          description: 已存在的表单实例ID
          example: 888
        formKey:
          type: string
          nullable: true
          description: 表单Key；与 formData 配合时自动创建实例
          example: travel_request
        formData:
          type: object
          additionalProperties: true
          nullable: true
          description: 表单数据（键值对）
          example:
            destination: 上海
            budget: 3200
        processKey:
          type: string
          description: approvalWorkflow/approvalCountersign/approvalOrSign/approvalSequential
          example: approvalWorkflow
        variables:
          type: object
          additionalProperties: true
          description: 流程变量
          example:
            amount: 3200
            reason: 客户拜访
        countersignUsers:
          type: array
          description: 会签用户ID列表（字符串）
          example: ["1002", "1003"]
          items:
            type: string
        countersignMode:
          type: string
          description: ALL/MAJORITY/OR
          example: ALL
        passRatio:
          type: number
          format: double
          description: 多数通过阈值（0~1）
          example: 1
    StartProcessResponse:
      type: object
      properties:
        processInstanceId:
          type: string
          description: 启动后的流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        message:
          type: string
          description: 返回消息
          example: Process started successfully
    ClaimTaskRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: 认领人用户ID（字符串）
          example: "1002"
    CompleteTaskRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: 办理人用户ID（字符串）
          example: "1002"
        approvalResult:
          type: string
          description: APPROVE/REJECT
          example: APPROVE
        comments:
          type: string
          description: 审批意见
          example: 同意，预算合理
    DelegateTaskRequest:
      type: object
      required: [userId, delegateUserId]
      properties:
        userId:
          type: string
          description: 当前处理人用户ID（字符串）
          example: "1002"
        delegateUserId:
          type: string
          description: 被委派人用户ID（字符串）
          example: "1005"
        comment:
          type: string
          description: 委派说明
          example: 我出差中，委派处理
    ResolveTaskRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: 被委派处理人用户ID（字符串）
          example: "1005"
        approvalResult:
          type: string
          description: 处理结果（APPROVE/REJECT）
          example: APPROVE
        comment:
          type: string
          description: 处理意见
          example: 已核实，同意
    ReassignTaskRequest:
      type: object
      required: [userId, newAssigneeId]
      properties:
        userId:
          type: string
          description: 发起转办用户ID（字符串）
          example: "1"
        newAssigneeId:
          type: string
          description: 新负责人用户ID（字符串）
          example: "1006"
        comment:
          type: string
          description: 转办说明
          example: 按组织调整转办
    ReturnTaskRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: 发起回退用户ID（字符串）
          example: "1002"
        comment:
          type: string
          description: 回退原因
          example: 信息不完整，退回补充
    ReturnToTargetRequest:
      type: object
      required: [userId, targetActivityId]
      properties:
        userId:
          type: string
          description: 发起回退用户ID（字符串）
          example: "1002"
        targetActivityId:
          type: string
          description: 目标节点 activityId
          example: managerApproveTask
        comment:
          type: string
          description: 回退说明
          example: 回退到部门主管节点
    CancelProcessRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: 发起撤销用户ID（字符串）
          example: "1001"
        comment:
          type: string
          description: 撤销原因
          example: 申请内容有误，撤销后重提
    TaskInfo:
      type: object
      properties:
        taskId:
          type: string
          description: 任务ID
          example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
        taskName:
          type: string
          description: 任务名称
          example: 部门主管审批
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        assignee:
          type: string
          nullable: true
          description: 当前处理人
          example: "1002"
        createTime:
          type: string
          format: date-time
          description: 任务创建时间
          example: "2026-02-24T10:20:30Z"
...
    BizRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 主键ID
          example: 1
        businessKey:
          type: string
          description: 业务主键
          example: BR-20260224-0001
        processInstanceId:
          type: string
          description: 流程实例ID
          example: "4b8d6c88-f1ab-11ee-9f9a-0242ac120002"
        processDefinitionId:
          type: string
          description: 流程定义ID
          example: approvalWorkflow:3:7504
        formInstanceId:
          type: integer
          format: int64
          nullable: true
          description: 表单实例ID
          example: 88
        applicantId:
          type: integer
          format: int64
          description: 申请人ID
          example: 1001
        applicantDeptId:
          type: integer
          format: int64
          nullable: true
          description: 申请人部门ID
          example: 10
        applicantPostId:
          type: integer
          format: int64
          nullable: true
          description: 申请人岗位ID
          example: 101
        title:
          type: string
          description: 申请标题
          example: 差旅申请-上海
        status:
          type: integer
          description: 0草稿/1已提交/2审批中/3通过/4拒绝/5退回/6撤销
          example: 2
        currentTaskId:
          type: string
          nullable: true
          description: 当前任务ID
          example: "c9a4c8b2-f1ab-11ee-9f9a-0242ac120002"
        currentAssigneeId:
          type: integer
          format: int64
          nullable: true
          description: 当前处理人ID
          example: 1002
        submitTime:
          type: string
          format: date-time
          nullable: true
          description: 提交时间
          example: "2026-02-24T10:20:30Z"
        finishTime:
          type: string
          format: date-time
          nullable: true
          description: 完成时间
          example: "2026-02-24T11:20:30Z"
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2026-02-24T10:20:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: "2026-02-24T10:30:00Z"
        isDeleted:
          type: integer
          description: 逻辑删除标记（0否/1是）
          example: 0
...